refactor(app:asctb-reporter): :recycle: minor refactors, convert into standalone part#1
