+++ apps/asctb-reporter/src/app/app.component.ts
@@ -2,10 +2,17 @@ import { Component, OnInit, inject } from '@angular/core';
 import { MatIconRegistry } from '@angular/material/icon';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { DomSanitizer } from '@angular/platform-browser';
-import { NavigationEnd, Router } from '@angular/router';
+import { NavigationEnd, Router, RouterModule } from '@angular/router';
 import { environment } from '../environments/environment';
+import { FileUploadModule } from './components/file-upload/file-upload.module';
+import { FooterModule } from './components/footer/footer.module';
+import { OrganTableSelectorModule } from './components/organ-table-selector/organ-table-selector.module';
 import { TrackingPopupComponent } from './components/tracking-popup/tracking-popup.component';
+import { TrackingPopupModule } from './components/tracking-popup/tracking-popup.module';
+import { DocsModule } from './modules/docs/docs.module';
+import { RootModule } from './modules/root/root.module';
 import { ConsentService } from './services/consent.service';
+import { MousePositionTrackerModule } from './services/mouse-position-tracker.module';
 
 declare let gtag: (arg1?: unknown, arg2?: unknown, arg3?: unknown) => void;
 
@@ -13,7 +20,16 @@ declare let gtag: (arg1?: unknown, arg2?: unknown, arg3?: unknown) => void;
   selector: 'app-reporter',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
-  standalone: false,
+  imports: [
+    DocsModule,
+    RootModule,
+    FileUploadModule,
+    OrganTableSelectorModule,
+    FooterModule,
+    TrackingPopupModule,
+    MousePositionTrackerModule,
+    RouterModule,
+  ],
 })
 export class AppComponent implements OnInit {
   readonly consentService = inject(ConsentService);
