+++ apps/asctb-reporter/src/app/components/controls/vis-controls.component.scss
@@ -54,40 +54,6 @@
   justify-content: space-between;
 }
 
-::ng-deep .mat-mdc-slide-toggle .mdc-switch.mdc-switch--selected:enabled .mdc-switch__icon {
-  fill: none !important;
-}
-
-::ng-deep .mat-mdc-slide-toggle .mdc-switch.mdc-switch--unselected:enabled .mdc-switch__icon {
-  fill: none !important;
-}
-
-.mat-mdc-slider .mdc-slider__track--active {
-  height: 0.125rem;
-}
-.mat-mdc-slider .mdc-slider__track--active {
-  height: 0.125rem;
-}
-
-mat-slider {
-  --mdc-slider-inactive-track-height: 0.18rem;
-  --mdc-slider-active-track-height: 0.18rem;
-  ::ng-deep .mdc-slider__thumb-knob {
-    border-width: 0.438rem;
-    height: 0.813rem;
-    width: 0.813rem;
-  }
-}
-
-::ng-deep .mat-mdc-slide-toggle {
-  --mdc-switch-unselected-handle-color: #ffffff;
-  --mdc-switch-unselected-focus-handle-color: #ffffff;
-  --mdc-switch-unselected-hover-handle-color: #ffffff;
-  --mdc-switch-unselected-track-color: #989898;
-  --mdc-switch-unselected-hover-track-color: #989898;
-  --mdc-switch-unselected-focus-track-color: #989898;
-}
-
 .vis-controls {
   font-size: 0.9rem;
 }
