+++ apps/asctb-reporter/src/app/components/legend/legend.component.scss
@@ -4,20 +4,6 @@
   box-shadow: none;
 }
 
-.mat-expansion-panel-header:not(.compare) {
-  border-top-left-radius: 0.5rem !important;
-  border-top-right-radius: 0.5rem !important;
-}
-
-.mat-expansion-panel-header {
-  transition: all 0.3s;
-  background-color: $secondary !important;
-}
-
-.mat-expansion-panel-header:hover {
-  background-color: darken($color: $secondary, $amount: 8%) !important;
-}
-
 .circle {
   height: 0.9375rem;
   width: 0.9375rem;
