+++ apps/asctb-reporter/src/app/modules/control-pane/control-pane.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { RouterModule } from '@angular/router';
 import { ControlsModule } from '../../components/controls/controls.module';
-import { LegendModule } from '../../components/legend/legend.module';
+import { LegendComponent } from '../../components/legend/legend.component';
 import { OmapControlsModule } from '../../components/omap-controls/omap-controls.module';
 import { SidenavModule } from '../../components/sidenav/sidenav.module';
 import { FunctionsModule } from '../functions/functions.module';
@@ -15,11 +15,11 @@ import { ControlPaneComponent } from './control-pane.component';
     CommonModule,
     SidenavModule,
     MatButtonModule,
-    LegendModule,
     FunctionsModule,
     ControlsModule,
     RouterModule,
     OmapControlsModule,
+    LegendComponent,
   ],
   exports: [ControlPaneComponent],
 })
