+++ apps/asctb-reporter/src/app/modules/root/root.component.scss
@@ -6,7 +6,7 @@
 }
 
 .control-pane {
-  width: 15.625rem;
+  width: 20.5rem;
   border-right: 0.1875rem solid #444a6510;
   min-width: 3.125rem;
 }
