+++ apps/asctb-reporter/src/app/modules/root/root.module.ts
@@ -1,24 +1,24 @@
-import { NgModule } from '@angular/core';
+import { DragDropModule } from '@angular/cdk/drag-drop';
 import { CommonModule } from '@angular/common';
-import { RootComponent } from './root.component';
+import { NgModule } from '@angular/core';
+import { MatBottomSheetModule } from '@angular/material/bottom-sheet';
+import { MatDialogModule } from '@angular/material/dialog';
 import { MatSidenavModule } from '@angular/material/sidenav';
-import { ControlPaneModule } from '../control-pane/control-pane.module';
-import { ReportModule } from '../../components/report/report.module';
-import { IndentedListModule } from '../../components/indented-list/indented-list.module';
-import { DebugLogsModule } from '../../components/debug-logs/debug-logs.module';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { CompareModule } from '../../components/compare/compare.module';
-import { NavbarModule } from '../navbar/navbar.module';
+import { DebugLogsModule } from '../../components/debug-logs/debug-logs.module';
+import { DoiModule } from '../../components/doi/doi.module';
 import { ErrorModule } from '../../components/error/error.module';
-import { LegendModule } from '../../components/legend/legend.module';
+import { IndentedListModule } from '../../components/indented-list/indented-list.module';
+import { InfoModule } from '../../components/info/info.module';
+import { LegendComponent } from '../../components/legend/legend.component';
+import { ReportModule } from '../../components/report/report.module';
+import { TableNestedMenuModule } from '../../components/table-nested-menu/table-nested-menu.module';
+import { ControlPaneModule } from '../control-pane/control-pane.module';
+import { NavbarModule } from '../navbar/navbar.module';
 import { PlaygroundModule } from '../playground/playground.module';
 import { TreeModule } from '../tree/tree.module';
-import { MatDialogModule } from '@angular/material/dialog';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { MatBottomSheetModule } from '@angular/material/bottom-sheet';
-import { DragDropModule } from '@angular/cdk/drag-drop';
-import { DoiModule } from '../../components/doi/doi.module';
-import { InfoModule } from '../../components/info/info.module';
-import { TableNestedMenuModule } from '../../components/table-nested-menu/table-nested-menu.module';
+import { RootComponent } from './root.component';
 
 @NgModule({
   declarations: [RootComponent],
@@ -32,7 +32,6 @@ import { TableNestedMenuModule } from '../../components/table-nested-menu/table-
     CompareModule,
     NavbarModule,
     ErrorModule,
-    LegendModule,
     PlaygroundModule,
     TreeModule,
     MatDialogModule,
@@ -42,6 +41,8 @@ import { TableNestedMenuModule } from '../../components/table-nested-menu/table-
     DoiModule,
     InfoModule,
     TableNestedMenuModule,
+
+    LegendComponent,
   ],
   exports: [RootComponent],
 })
