+++ apps/asctb-reporter/src/styles.scss
@@ -1,8 +1,11 @@
+/** New Changes */
+@use 'light-theme';
+@use 'theming';
+@use 'utils';
+
 @import 'bootstrap/dist/css/bootstrap.css';
 @import './assets/stylesheets/colors';
-@import './theme.scss';
 @import 'https://fonts.googleapis.com/icon?family=Material+Icons';
-
 * {
   font-family: 'Inter', sans-serif;
   box-sizing: border-box;
@@ -12,122 +15,9 @@ html,
 body {
   height: 100%;
 }
-body {
-  margin: 0;
-  font-family: Roboto, 'Helvetica Neue', sans-serif;
-}
 
-.logo-link {
-  cursor: pointer;
-}
+/** New Changes */
 
-button:focus {
-  outline: 0 !important;
-}
-
-/* SCROLL BAR */
-::-webkit-scrollbar {
-  width: 0.125rem;
-  height: 0.3125rem;
-  backdrop-filter: blur(0.625rem);
-}
-
-::-webkit-scrollbar-track {
-  background-color: rgba(255, 255, 255, 0.089);
-}
-
-::-webkit-scrollbar-thumb {
-  background-color: #444a56a8;
-  border-radius: 0.625rem;
-}
-
-.secondary {
-  background-color: $secondary !important;
-}
-
-.select-border {
-  border: 0.125rem solid $secondary !important;
-}
-
-/* SNACKBAR */
-.success-snackbar {
-  background: #4caf50;
-}
-
-.mat-simple-snackbar-action {
-  color: white;
-}
-
-.error-snackbar {
-  background: rgb(194, 87, 87);
-}
-
-.warning-snackbar {
-  background: #ffa500;
-}
-
-/* EXPANSION PANEL */
-.mepNoPadding .mat-expansion-panel-body {
-  padding: 0.9375rem !important;
-  border: 0.125rem solid $secondary;
-  border-bottom-left-radius: 0.5rem !important;
-  border-bottom-right-radius: 0.5rem !important;
-}
-
-/* CARD STYLES */
-.card {
-  border: none;
-  border-radius: 0;
-}
-
-.card-header {
-  background-color: transparent;
-  border-bottom: 0;
-}
-
-.card-footer {
-  border-top: 0;
-  background-color: transparent;
-}
-
-.card-body {
-  overflow-y: auto;
-}
-
-// BOTTOM SHEET
-
-.bottom-sheet-style {
-  border: 0.1875rem solid #ececec;
-  max-width: 18.75rem !important;
-  box-shadow: 0px 5px 15px 10px rgb(0 0 0 / 5%);
-  border-top-right-radius: 0.5rem;
-  border-top-left-radius: 0.5rem;
-  padding: 0;
-}
-
-// SIDE NAV
-
-::ng-deep .mat-drawer-backdrop .ng-star-inserted .mat-drawer-shown {
-  background-color: green !important;
-}
-
-h1,
-h2,
-h3,
-h4 {
-  font-weight: 600;
-}
-
-markdown .md-img {
-  border-radius: 5px;
-  width: 100%;
-  box-shadow: 0px 0px 15px 10px #7272720a;
-}
-
-markdown strong {
-  font-weight: 600;
-}
-
-.mat-snack-bar-container {
-  min-width: fit-content !important;
+@include utils.global-styles() {
+  @include theming.theme(light-theme.$theme);
 }
